Refactor server service implmentation to own module
